clients cleanup
