+++ scrapper/src/main/java/edu/java/client/StackoverflowWebClient.java
@@ -1,19 +1,18 @@
 package edu.java.client;
 
 import edu.java.dto.StackoverflowQuestionResponse;
-import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Mono;
 
 @Component
-public class StackoverflowClient {
+public class StackoverflowWebClient {
 
     private static final String URL = "/questions/%d?order=%s&sort=%s&site=stackoverflow";
     private final WebClient webClient;
 
-    public StackoverflowClient(@Qualifier("stackoverflow") WebClient webClient) {
-        this.webClient = webClient;
+    public StackoverflowWebClient(WebClient stackoverflowClient) {
+        this.webClient = stackoverflowClient;
     }
 
     public Mono<StackoverflowQuestionResponse> fetchQuestion(long questionId, String sort, String order) {
