+++ scrapper/src/test/java/edu/java/scrapper/GithubWebClientTest.java
@@ -2,8 +2,7 @@ package edu.java.scrapper;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.client.WireMock;
-import edu.java.client.StackoverflowClient;
-import java.time.OffsetDateTime;
+import edu.java.client.GithubWebClient;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
@@ -11,7 +10,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.web.reactive.function.client.WebClient;
 import reactor.test.StepVerifier;
 
-public class StackoverflowClientTest {
+public class GithubWebClientTest {
     private static WireMockServer wireMockServer;
 
     @BeforeAll
@@ -28,37 +27,33 @@ public class StackoverflowClientTest {
 
     @Test
     @DisplayName("test for check the required response body")
-    public void testFetchQuestion() {
+    public void testFetchRepository() {
         // Arrange
-        long questionId = 123456;
-        String order = "activity";
-        String sort = "desc";
-        OffsetDateTime fixedTime = OffsetDateTime.parse("2022-02-21T12:34:56Z");
+        String owner = "testOwner";
+        String repo = "testRepo";
 
-        wireMockServer.stubFor(WireMock.get(WireMock.urlPathEqualTo("/questions/123456"))
+        wireMockServer.stubFor(WireMock.get(WireMock.urlPathEqualTo("/repos/testOwner/testRepo"))
             .willReturn(WireMock.aResponse()
                 .withStatus(200)
                 .withHeader("Content-Type", "application/json")
                 .withBody(
-                    "{\"question_id\":123456,\"title\":\"Test Question\",\"link\":\"https://stackoverflow.com/q/123456\",\"last_activity_date\":\"" +
-                        fixedTime + "\"}")
+                    "{\"name\":\"testRepo\",\"full_name\":\"testOwner/testRepo\",\"owner\":\"testOwner\",\"description\":\"Test Repo\",\"html_url\":\"https://github.com/testOwner/testRepo\"}")
             ));
 
         // Act
         WebClient webClient = WebClient.builder().baseUrl("http://localhost:" + wireMockServer.port()).build();
-        StackoverflowClient stackOverflowClient = new StackoverflowClient(webClient);
+        GithubWebClient gitHubWebClient = new GithubWebClient(webClient);
 
         // Assert
-        StepVerifier.create(stackOverflowClient.fetchQuestion(questionId, sort, order))
-            // Then
+        StepVerifier.create(gitHubWebClient.fetchRepository(owner, repo))
             .expectNextMatches(response ->
-                response.getQuestionId() == 123456 &&
-                    response.getTitle().equals("Test Question") &&
-                    response.getLink().equals("https://stackoverflow.com/q/123456") &&
-                    response.getLastActivityDate().equals(fixedTime)
+                response.getName().equals("testRepo") &&
+                    response.getFullName().equals("testOwner/testRepo") &&
+                    response.getOwner().equals("testOwner") &&
+                    response.getDescription().equals("Test Repo") &&
+                    response.getHtmlUrl().equals("https://github.com/testOwner/testRepo")
             )
             .expectComplete()
             .verify();
     }
-
 }
